Move changes to dummy_data into dummy_data_nextgen
