+++ ehrql/dummy_data/query_info.py
@@ -1,7 +1,6 @@
 import dataclasses
 from collections import defaultdict
-from datetime import date, timedelta
-from functools import cached_property, lru_cache
+from functools import cached_property
 
 from ehrql.query_model.introspection import all_unique_nodes, get_table_nodes
 from ehrql.query_model.nodes import (
@@ -15,7 +14,6 @@ from ehrql.query_model.nodes import (
     Value,
     get_root_frame,
 )
-from ehrql.query_model.table_schema import Constraint
 
 
 @dataclasses.dataclass
@@ -30,17 +28,11 @@ class ColumnInfo:
     _values_used: set = dataclasses.field(default_factory=set)
 
     @classmethod
-    def from_column(cls, name, column, extra_constraints=()):
+    def from_column(cls, name, column):
         type_ = column.type_
         if hasattr(type_, "_primitive_type"):
             type_ = type_._primitive_type()
-        return cls(
-            name,
-            type_,
-            constraints=normalize_constraints(
-                tuple(column.constraints) + tuple(extra_constraints)
-            ),
-        )
+        return cls(name, type_, constraints=tuple(column.constraints))
 
     def __post_init__(self):
         self._constraints_by_type = {type(c): c for c in self.constraints}
@@ -109,10 +101,6 @@ class QueryInfo:
         all_nodes = all_unique_nodes(*variable_definitions.values())
         by_type = get_nodes_by_type(all_nodes)
 
-        extra_constraints = query_to_column_constraints(
-            variable_definitions["population"]
-        )
-
         tables = {
             # Create a TableInfo object …
             table.name: TableInfo.from_table(table)
@@ -140,9 +128,7 @@ class QueryInfo:
             if column_info is None:
                 # … insert a ColumnInfo object into the appropriate table
                 column_info = ColumnInfo.from_column(
-                    name,
-                    table.schema.get_column(name),
-                    extra_constraints=extra_constraints.get(column, ()),
+                    name, table.schema.get_column(name)
                 )
                 table_info.columns[name] = column_info
             # Record the ColumnInfo object associated with each SelectColumn node
@@ -190,164 +176,3 @@ def get_nodes_by_type(nodes):
 
 def sort_by_name(iterable):
     return sorted(iterable, key=lambda i: i.name)
-
-
-@lru_cache
-def query_to_column_constraints(query):
-    """Converts a query (typically a population definition) into
-    constraints that would have to be applied to a record in order
-    to satisfy it."""
-    match query:
-        case Function.And(lhs=lhs, rhs=rhs):
-            left = query_to_column_constraints(lhs)
-            right = query_to_column_constraints(rhs)
-            keys = set(left) | set(right)
-            return {k: left.get(k, []) + right.get(k, []) for k in keys}
-        case Function.Or(lhs=lhs, rhs=rhs):
-            left = query_to_column_constraints(lhs)
-            right = query_to_column_constraints(rhs)
-            result = {}
-            for k, v in left.items():
-                try:
-                    result[k] = list(set(v) & set(right[k]))
-                except KeyError:
-                    pass
-            for k, v in list(result.items()):
-                if not v:
-                    del result[k]
-            return result
-        case Function.EQ(
-            lhs=SelectColumn() as lhs,
-            rhs=Value(value=value),
-        ):
-            return {lhs: [Constraint.Categorical(values=(value,))]}
-        case Function.EQ(
-            lhs=Function.YearFromDate(source=SelectColumn() as column),
-            rhs=Value(value=year),
-        ):
-            return {
-                column: [
-                    Constraint.GeneralRange(
-                        minimum=date(year, 1, 1),
-                        maximum=date(year, 12, 31),
-                    )
-                ]
-            }
-        case Function.In(
-            lhs=SelectColumn() as lhs,
-            rhs=Value(value=values),
-        ):
-            return {lhs: [Constraint.Categorical(values=values)]}
-        case Function.GE(
-            lhs=Function.DateDifferenceInYears(
-                lhs=Value(value=reference_date), rhs=column
-            ),
-            rhs=Value(value=difference),
-        ):
-            return {
-                column: [
-                    Constraint.GeneralRange(
-                        maximum=reference_date - timedelta(days=365 * difference)
-                    )
-                ]
-            }
-        case Function.LE(
-            lhs=Function.DateDifferenceInYears(
-                lhs=Value(value=reference_date), rhs=column
-            ),
-            rhs=Value(value=difference),
-        ):
-            return {
-                column: [
-                    Constraint.GeneralRange(
-                        minimum=reference_date.replace(
-                            year=reference_date.year - difference
-                        )
-                    )
-                ]
-            }
-        case Function.LT(
-            lhs=Function.DateAddYears(
-                lhs=SelectColumn() as column,
-                rhs=Value(value=difference),
-            ),
-            rhs=Value(value=reference_date),
-        ):
-            return {
-                column: [
-                    Constraint.GeneralRange(
-                        maximum=reference_date.replace(
-                            year=reference_date.year - difference
-                        ),
-                        includes_maximum=False,
-                    )
-                ]
-            }
-        case Function.GT(lhs=SelectColumn() as column, rhs=Value(value=min_value)):
-            return {
-                column: [
-                    Constraint.GeneralRange(minimum=min_value, includes_minimum=False)
-                ]
-            }
-        case Function.GE(lhs=SelectColumn() as column, rhs=Value(value=min_value)):
-            return {
-                column: [
-                    Constraint.GeneralRange(minimum=min_value, includes_minimum=True)
-                ]
-            }
-        case Function.LT(lhs=SelectColumn() as column, rhs=Value(value=max_value)):
-            return {
-                column: [
-                    Constraint.GeneralRange(maximum=max_value, includes_maximum=False)
-                ]
-            }
-        case Function.LE(lhs=SelectColumn() as column, rhs=Value(value=max_value)):
-            return {
-                column: [
-                    Constraint.GeneralRange(maximum=max_value, includes_maximum=True)
-                ]
-            }
-        case Function.IsNull(source=SelectColumn() as column):
-            return {column: [Constraint.NotNull()]}
-
-    return {}
-
-
-def normalize_constraints(constraints):
-    group_by_type = defaultdict(list)
-    for constraint in constraints:
-        group_by_type[type(constraint)].append(constraint)
-    if len(group_by_type[Constraint.Categorical]) > 1:
-        constraint, *rest = group_by_type[Constraint.Categorical]
-        for more in rest:
-            constraint = Constraint.Categorical(
-                values=set(constraint.values) & set(more.values)
-            )
-        group_by_type[Constraint.Categorical] = [constraint]
-    if len(ranges := group_by_type[Constraint.GeneralRange]) > 1:
-        minimum = None
-        maximum = None
-        for r in ranges:
-            if minimum is None:
-                minimum = r.minimum
-            elif r.minimum is not None:
-                minimum = max(minimum, r.minimum)
-            if maximum is None:
-                maximum = r.maximum
-            elif r.maximum is not None:
-                maximum = min(maximum, r.maximum)
-
-        includes_minimum = minimum is None or all(r.validate(minimum) for r in ranges)
-        includes_maximum = maximum is None or all(r.validate(maximum) for r in ranges)
-        group_by_type[Constraint.GeneralRange] = [
-            Constraint.GeneralRange(
-                minimum=minimum,
-                maximum=maximum,
-                includes_maximum=includes_maximum,
-                includes_minimum=includes_minimum,
-            )
-        ]
-
-    return tuple(
-        [constraint for group in group_by_type.values() for constraint in group]
-    )
